+++ lib/flutter_ozow.dart
@@ -1,4 +1,0 @@
-library flutter_ozow;
-
-export 'res/ozow_payment.dart';
-export 'res/ozow_payment_ui.dart';